requirements.txt
# ── Web framework ────────────────────────────────────────────────────────────
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# ── Validation / Settings ─────────────────────────────────────────────────────
pydantic==2.6.1
pydantic-settings==2.2.0
python-dotenv==1.0.1

# ── Auth & Security ───────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.2

# ── Background tasks ──────────────────────────────────────────────────────────
celery==5.3.6
redis==5.0.1
APScheduler==3.10.4

# ── Document parsing ──────────────────────────────────────────────────────────
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.2.0
watchdog==3.0.0
aiofiles==23.2.1

# ── Matching / NLP ────────────────────────────────────────────────────────────
rapidfuzz==3.6.1

# ── Reporting ─────────────────────────────────────────────────────────────────
openpyxl==3.1.2
reportlab==4.1.0

# ── External integrations ─────────────────────────────────────────────────────
msal==1.26.0
httpx==0.26.0
playwright==1.41.2

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dateutil==2.8.2
pytz==2024.1
tenacity==8.2.3
structlog==24.1.0
